Trigger creation of inputs/ and outputs/ directories
Git doesn't track directories, it really tracks lines of files.  As a result
to add a directory to a git repository you have to create the directory, add
a small file to it, and then add the file.  Git then tracks the file inside
that directory.  This IS odd the first time you come across it, but that's
the reason why this commit is called "trigger creation of inputs and outputs
directories" rather than "Add README.md files to inputs and outputs".

After all the intent is to create to directories to allow you to put (eg)
CSV file data in inputs and the output data in outputs.

Michael.