Add initial support for energy tracking
This is split into 3 pieces:

* The code for accessing RAPL under linux (LinuxRAPL.cpp).  This gives
  energy (essentially) in (micro) joules.

* Tracker code which keeps track of energy and time, and assists with
  figuring out things like average power (in watts)

* The includes to hold all this together.

This needs refining and possibly redoing, and definitely needs linting and
restructuring.

It also needs validating against external PSU monitoring because I suspect
this is double counting at the moment. It's a start though.

This has been put together rapidly and inspired by a number of sources
including CodeCarbon and a piece of code called CPPJoules which briefly
existed and then did a disappearing act.

As a result, should this energy measuring code prove useful it will very
likely get extracted out as a separate library.  It's creation and inclusion
here is to support the "Carbon Tracking" element of the students' project.

Things that need doing:

* Checking for double counting (common issue)
* Checking with external PSU
* Linting
* Documenting
* The warnings generated because of the library need squishing.