Add basic energy tracking to the code
This is intended as a test for the energy tracking code added in this branch
but also to provides an example of how this can be integrated into the
codebase.  It's very much version 0 of such an example, and like the energy
tracking code, it needs a bit of work.  It's loosely based on the approach
that CodeCarbon takes in python, but can't take exactly the same approach
due to differences between C++ and python.  (The latter has decorators)

This creates a tracker to measure the data, and a data structure
to accumulate results from the tracker. It adds a preupdate system
to start tracking energy usage before the gravity calculations happen
and then a postupdate system to stop the tracker and capture the results.

So the process is:
* Create tracker, zero store
* When iterating
  - PreUpdate System - start tracking
  - OnUpdate - do calculations
  - PostUpdate System - stop tracking
    .. and update our store about the system
* On exit report the results