Switched energy tracking over to a single header file
Single header file based libraries are a relatively common thing at the
moment in the C++ world - since someone knows that in order to use the file
all they need to do is pull in one file and everything works.

This refactoring is a 'make it right' piece of work after the previous 'make it work'
phase of work already integrated.